:recycle: Set up API v2
