+++ .env.sample
@@ -4,6 +4,13 @@ JWT_ISSUER=
 JWT_VALIDITY=
 JWT_VALIDITY_FOR_API=
 
+HASURA_JWT_SECRET_KEY=
+HASURA_JWT_HMAC_SHA_ALGO=
+HASURA_JWT_ISSUER=
+HASURA_JWT_VALIDITY=
+HASURA_JWT_ISSUER=
+HASURA_JWT_AUDIENCE=
+
 OTP_CODE_VALIDITY=
 
 CLOUDINARY_CLOUD_NAME=
