+++ config/routes.rb
@@ -5,6 +5,11 @@ Rails.application.routes.draw do
   root :controller => 'static', :action => '/'
 
   namespace "api" do
+    namespace "v2", defaults: { format: "json" } do
+      post "auth/send_pin", to: "authentication#send_pin"
+      post "auth/hasura", to: "authentication#hasura"
+    end
+
     namespace "v1", defaults: { format: "json" } do
 
       get "/browse/fetch_packages", to: "packages#index" #temporary redirect for old browse apps
