+++ config/secrets.yml
@@ -16,6 +16,12 @@ base: &BASE
     issuer: <%= ENV['JWT_ISSUER'] %>
     validity: <%= ENV['JWT_VALIDITY'] %>
     validity_for_api: <%= ENV['JWT_VALIDITY_FOR_API']%>
+  hasura:
+    secret_key: <%= ENV['HASURA_JWT_SECRET_KEY'] %>
+    hmac_sha_algo: <%= ENV['HASURA_JWT_HMAC_SHA_ALGO'] %>
+    issuer: <%= ENV['HASURA_JWT_ISSUER'] %>
+    audience: <%= ENV['HASURA_JWT_AUDIENCE'] %>
+    validity: <%= ENV['HASURA_JWT_VALIDITY'] %>
   token:
     otp_code_validity: <%= ENV['OTP_CODE_VALIDITY'] %>
   cloudinary:
