+++ spec/support/env.rb
@@ -14,6 +14,12 @@ ENV['CLOUDINARY_API_SECRET'] = "test"
 ENV['PREVENT_CLOUDINARY_IMAGE_DELETION'] = "false"
 ENV['SECRET_KEY_BASE'] = "21ff5a7a2ed8a8f5ac06c36a5118055bbdd7a7268907328d04073e9e8879c9f0636d031ff06b65befa41e74df45835c985df735d64411fa80ddd64a4303d0c9b"
 ENV['SECRET_KEY'] = "123eadadaf"
+ENV['HASURA_JWT_SECRET'] = "0246d73b8a1cf-dfa7f075-f2f0-4f14-b1bc-5b08a3fc6513"
+ENV['HASURA_JWT_HMAC_SHA_ALGO'] = "HS256"
+ENV['HASURA_JWT_ISSUER'] = "goodcity"
+ENV['HASURA_JWT_AUDIENCE'] = "goodcity"
+ENV['HASURA_JWT_VALIDITY'] = "3600"
+
 ENV['JWT_SECRET_KEY'] = "0246d73b8a1cf"
 ENV['JWT_HMAC_SHA_ALGO'] = "HS256"
 ENV['JWT_ISSUER'] = "GoodCityTest"
